+++ denode/index.js
@@ -2,9 +2,11 @@
 
 const Path = require('path')
 const Electron = require('electron')
+const STDIN = []
+let win
 
 Electron.app.on('ready', function() {
-  var win = new Electron.BrowserWindow({ width: 800, height: 600, title: 'DeNode' });
+  win = new Electron.BrowserWindow({ width: 800, height: 600, title: 'DeNode' });
   win.loadURL('file://' + Path.join(__dirname, 'index.html') + '?app=' + encodeURIComponent(process.argv[2] || ''));
   win.on('closed', function() {
     Electron.app.quit()
@@ -20,6 +22,12 @@ Electron.app.on('ready', function() {
       })
     })
   })
+
+  win.webContents.on('dom-ready', function() {
+    for (let i = 0, length = STDIN.length; i < length; ++i) {
+      win.webContents.send('stdin', STDIN[i])
+    }
+  })
 })
 
 Electron.ipcMain.on('stdout', function(_, data) {
@@ -28,3 +36,11 @@ Electron.ipcMain.on('stdout', function(_, data) {
 Electron.ipcMain.on('stderr', function(_, data) {
   process.stderr.write(data)
 })
+process.stdin.on('data', function(chunk) {
+  const stringish = chunk.toString()
+  if (win) {
+    win.webContents.send('stdin', stringish)
+  }
+  STDIN.push(stringish)
+})
+process.stdin.resume()